feat: finish products repository and controller

+++ GeekShopping.productAPI/Repository/IProductRepository.cs
@@ -6,12 +6,12 @@ namespace GeekShopping.productAPI.Repository
     {
         Task<IEnumerable<ProductVO>> FindAll();
 
-        Task<ProductVO> FindById(int id);
+        Task<ProductVO> FindById(long id);
 
         Task<ProductVO> Create(ProductVO vo);
 
         Task<ProductVO> Update(ProductVO vo);
 
-        Task<bool> Delete(int id);
+        Task<bool> Delete(long id);
     }
 }
